requirements.txt
# Core FastAPI
fastapi[standard]==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Keycloak Integration
PyJWT==2.10.1
cryptography==44.0.0

# Google AI Integration
google-genai==1.0.0
google-auth==2.36.0

# WebSocket & Background Tasks
websockets==14.1
python-socketio==5.11.4
celery==5.4.0
redis==5.2.1
aiojobs==1.3.0

# HTTP Client
httpx==0.28.1

# Utilities
python-dateutil==2.9.0.post0
email-validator==2.2.0
typer==0.15.1

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Production
gunicorn==23.0.0

# Monitoring & Logging
structlog==24.4.0
sentry-sdk==2.20.0
